+++ src/main/java/jdistlib/math/Bessel.java
@@ -195,8 +195,8 @@ public class Bessel {
 		if (na < 0) {
 			/* Using Abramowitz & Stegun  9.1.2
 			 * this may not be quite optimal (CPU and accuracy wise) */
-			return (alpha - na == 0.5) ? 0 : j(x, -alpha) * cospi(alpha) + (alpha == na ? 0 :
-				y(x, -alpha) * sinpi(alpha)); // PR#15554 fix
+			return j(x, -alpha) * cospi(alpha) + (alpha == na ? 0 :
+				y(x, -alpha) * sinpi(alpha));
 		}
 		int nb = 1 + na;
 		alpha -= (nb - 1);
@@ -249,8 +249,8 @@ public class Bessel {
 		if (na < 0) {
 			/* Using Abramowitz & Stegun  9.1.2
 			 * this may not be quite optimal (CPU and accuracy wise) */
-			return (alpha - na == 0.5) ? 0 : y(x, -alpha) * cospi(alpha) + (alpha == na ? 0 :
-				j(x, -alpha) * sinpi(alpha)); // PR#15554 fix
+			return y(x, -alpha) * cospi(alpha) - (alpha == na ? 0 :
+				j(x, -alpha) * sinpi(alpha));
 		}
 		int nb = 1 + na;
 		alpha -= (nb - 1);